refactor: 관심강좌 관련 기능은 BookmarkRepository 로 통합
