+++ app/src/main/java/com/wafflestudio/snutt2/data/lecture_search/LectureSearchRepository.kt
@@ -9,13 +9,10 @@ import com.wafflestudio.snutt2.domainmodel.SearchedLecture
 import com.wafflestudio.snutt2.lib.network.Result
 import com.wafflestudio.snutt2.lib.network.dto.core.LectureBuildingDto
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.StateFlow
 
 interface LectureSearchRepository {
 
     val recentSearchedDepartmentTags: Flow<List<SearchTag>>
-
-    val firstBookmarkAlert: StateFlow<Boolean>
 
     fun getLectureSearchResultStream(
         year: Long,
@@ -31,8 +28,6 @@ interface LectureSearchRepository {
     fun storeRecentSearchedDepartment(tag: SearchTag)
 
     fun removeRecentSearchedDepartment(tag: SearchTag)
-
-    fun setFirstBookmarkAlertShown()
 
     suspend fun getBuildings(places: List<String>): Result<List<LectureBuildingDto>>
